factor out css-related methods from util module
