+++ lib/app.js
@@ -10,6 +10,7 @@ var mkdirp = require('mkdirp');
 var rimraf = require('rimraf');
 var Progger = require('progger');
 var util = require(path.resolve(__dirname, './', 'util'));
+var css = require(path.resolve(__dirname, './', 'css'));
 var image = require(path.resolve(__dirname, './', 'image'));
 var msg = require(path.resolve(__dirname, './', 'msg'));
 var analytics = require(path.resolve(__dirname, './', 'analytics'));
@@ -280,10 +281,10 @@ module.exports = function (args, opts) {
       msg.log('info', 'generatingCss');
     }
 
-    return util.createCssRules(results, opts);
+    return css.createRules(results, opts);
   }).then(function (cssArray) {
 
-    var cssString = util.mungeCss(cssArray);
+    var cssString = css.munge(cssArray);
 
     if (opts.stdout) {
 
@@ -309,7 +310,7 @@ module.exports = function (args, opts) {
       msg.log('info', 'saveCss');
     }
 
-    return util.saveCss(path.resolve(outputDir, cssFilename), cssArray, opts);
+    return css.save(path.resolve(outputDir, cssFilename), cssArray, opts);
   }).then(function () {
 
     if (!opts.nopngdata) {
