+++ lib/util.js
@@ -5,7 +5,6 @@ var fs = Bluebird.promisifyAll(require('fs'));
 var path = require('path');
 var mime = require('mime');
 var _ = require('lodash');
-var pretty = require('cssbeautify');
 
 module.exports = {
 
@@ -36,85 +35,6 @@ module.exports = {
       var isSvg = mime.lookup(file) === 'image/svg+xml';
       return isFile && isSvg;
     });
-  },
-
-  // returns a string that can be saved as a CSS file
-  createCssRules: function createCssRules(results, opts) {
-    var cssClassnamePrefix = opts.classname !== null ? opts.classname : '';
-    var css = [];
-    var cssSvg = '';
-    var cssNoInlineSvg = '';
-    var cssNoDataUri = '';
-    if (!opts.killcomment) {
-      cssSvg += '/* generated by iconr */\n\n';
-    }
-
-    _.each(results, function (result) {
-      var height = !isNaN(result.height) ? 'height:' + result.height + 'px;' : '';
-      var width = !isNaN(result.width) ? 'width:' + result.width + 'px;' : '';
-      cssSvg += '.#' + (cssClassnamePrefix + result.name) + '{' + height + width + 'background-image:url("' + result.svgdatauri + '");}';
-      if (!opts.nopngdata) {
-        cssNoInlineSvg += '.no-inlinesvg .' + (cssClassnamePrefix + result.name) + '{background-image:url("' + result.pngdatauri + '");}';
-      }
-      if (!opts.nopng && !opts.stdout) {
-        cssNoDataUri += '.no-datauri .' + (cssClassnamePrefix + result.name) + '{background-image:url("' + result.pngpath + '");}';
-      }
-    });
-
-    css.push(cssSvg, cssNoInlineSvg, cssNoDataUri);
-
-    return css;
-  },
-
-  // returns a "beautified" version of a css string
-  prettyCss: function prettyCss(css) {
-    return pretty(css);
-  },
-
-  // merge the css array into a string
-  mungeCss: function mungeCss(cssArray) {
-    return cssArray.join('');
-  },
-
-  // saves CSS file(s) to disk
-  // can't use arrow syntax here because of "this" reference
-  saveCss: function saveCss(filename, cssArr, opts) {
-
-    // save separate css files
-    if (opts.separatecss) {
-
-      // prettify the CSS if necessary
-      var css = opts.pretty ? cssArr.map(function (cssStr) {
-        return pretty(cssStr);
-      }) : css;
-
-      // save CSS with SVG data URIs
-      fs.writeFileAsync(filename + '.css', css[0]);
-
-      // save CSS with fallback PNG data URIs
-      if (!opts.nopngdata && css[1].length !== 0) {
-        fs.writeFileAsync(filename + '-noinlinesvg.css', css[1]);
-      }
-
-      // save CSS with fallback PNG image paths
-      if (!opts.nopng && css[2].length !== 0) {
-        fs.writeFileAsync(filename + '-nodatauri.css', css[2]);
-      }
-    } else {
-
-      // combine CSS into string
-      var css = this.mungeCss(cssArr);
-
-      // prettify the CSS if necessary
-      if (opts.pretty) {
-        css = this.prettyCss(css);
-      }
-
-      // write CSS into a single file
-      fs.writeFileAsync(filename + '.css', css);
-    }
-
-    return;
   }
 
 };
